src/auth/requirements.txt
anyio==3.7.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
identify==2.5.24
idna==3.4
mysqlclient==2.1.1
nodeenv==1.8.0
platformdirs==3.5.1
pre-commit==3.3.2
protobuf==3.20.3
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
virtualenv==20.23.0
